+++ spec/acceptance/community_instance_spec.rb
@@ -20,6 +20,10 @@ describe 'mariadb class' do
     		ensure => 'present',
     	}
 
+      mysql::backup::mysqldump { 'test':
+        destination => '/backup',
+      }
+
       EOF
 
       # Run it twice and test for idempotency
@@ -49,6 +53,20 @@ describe 'mariadb class' do
     describe file("/etc/mysql/test/my.cnf") do
       it { should be_file }
       its(:content) { should match '[mysqld]' }
+    end
+
+    # /usr/local/bin/backupmysqldump
+    describe file("/usr/local/bin/backupmysqldump") do
+      it { should be_file }
+      its(:content) { should match 'puppet managed file' }
+    end
+
+    it "backupmysqldump" do
+      expect(shell("/usr/local/bin/backupmysqldump").exit_code).to be_zero
+    end
+
+    it "ls backups" do
+      expect(shell("ls /backup").exit_code).to be_zero
     end
 
   end