+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * added **sql_mode** and **lower_case_table_names** to **mysqld section**
 * fixed **unless** and **onlyif** for **mysql_sql**
+* moved backup scripts to files folder
 
 ## 0.4.12